+++ Packages/com.pg649.creaturegenerator/Runtime/Metaball/MeshGenerator.cs
@@ -31,6 +31,8 @@ namespace MarchingCubesProject
 
         private NormalRenderer normalRenderer;
 
+        private IRigSolver rigSolver;
+
         void Start() 
         {
         }
@@ -42,6 +44,7 @@ namespace MarchingCubesProject
         public void ApplySettings(MeshSettings settings, DebugSettings debugSettings)
         {
             mode = settings.Mode;
+            rigSolver = IRigSolver.Get(settings.RigSolver);
             material = settings.Material;
             smoothNormals = settings.SmoothNormals;
             drawNormals = debugSettings.DrawNormals;
@@ -182,7 +185,7 @@ namespace MarchingCubesProject
             }
             mesh.bindposes = bindPoses.ToArray();
 
-            mesh.boneWeights = BoneUtil.CalcBoneWeightsTheStupidWay(mesh, bones, transform);
+            mesh.boneWeights = rigSolver.CalcBoneWeights(mesh, bones, transform);
 
             go.AddComponent<MeshFilter>();
             go.AddComponent<MeshRenderer>();