+++ Packages/com.pg649.creaturegenerator/Runtime/SkeletonGeneration/Generators/BipedGenerator.cs
@@ -37,6 +37,8 @@ public class BipedGenerator {
         foreach (var arm in arms)
             armAttachmentBone.LinkChild(arm);
 
+        root.AttachmentHint.Offset = new(0, instance.LegLengths.Sum() + instance.HipLength * 0.5f + SkeletonAssembler.FootHeight, 0);
+
         return new SkeletonDefinition(root, new LimitTable(HumanoidJointLimits), instance);
     }
 
@@ -134,7 +136,7 @@ public class BipedGenerator {
             Category = BoneCategory.Foot,
             AttachmentHint = new AttachmentHint
             {
-                Offset = new Vector3(0.0f, 0.0f, -0.5f)
+                Position = new RelativePosition(0.0f, -1.0f, 1.05f),
             }
         };